+++ implementation/site/main.js
@@ -8,7 +8,8 @@ NProgress.configure({showSpinner: false});
         //fullScreen: true,
         template: '@site.html',
         contextRegion: 'center',
-        defaultContext: 'Home'
+        defaultContext: 'Home',
+        baseAjaxURI: ''
 
     }).run();